+++ examples/data-sources/pingfederate_authentication_api_application/data-source.tf
@@ -1,0 +1,3 @@
+data "pingfederate_authentication_api_application" "myAuthenticationApiApplicationExample" {
+	application_id = "example"
+}